% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNFCT.R
\name{nfct_1d}
\alias{nfct_1d}
\alias{nfct_adjoint_1d}
\alias{nfst_1d}
\alias{nfst_adjoint_1d}
\title{1-D Non-Uniform Fast Cosine/Sine Tranform}
\usage{
nfct_1d(x, f_hat)

nfct_adjoint_1d(x, f, N)

nfst_1d(x, f_hat)

nfst_adjoint_1d(x, f, N)
}
\arguments{
\item{x}{(real) vector of nodes of length \code{M}, must be in \code{[-0.5,0.5)}.}

\item{f_hat}{(complex) vector of \eqn{\hat{f}}{f_hat} entries -
(of length \code{N}) and length must be even.}

\item{f}{frequencies for adjoint, same length as \code{x}}

\item{N}{number of frequencies for transform, specified for adjoint.}
}
\description{
Similar to the NFFT, except the cosine/sine transform instead of the
FFT. Note that all values of \code{x} must be in \eqn{[0,1/2)}.
}
\section{Functions}{
\itemize{
\item \code{nfct_adjoint_1d}: The adjoint transform

\item \code{nfst_1d}: The sine transform

\item \code{nfst_adjoint_1d}: The adjoint sine transform
}}

\examples{

set.seed(20190722)
x <- runif(19) * 0.5
f_hat = runif(14)

nfct_1d(x, f_hat)
nfst_1d(x, f_hat)
f <- nfct_1d(x, f_hat)
g <- nfst_1d(x, f_hat)
nfct_adjoint_1d(x,f,14)
nfst_adjoint_1d(x,g,14) # note the length!
}
