% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRadon.R
\name{nfft_radon}
\alias{nfft_radon}
\alias{nfft_inv_radon}
\title{Radon transform using NFFT}
\usage{
nfft_radon(
  image,
  theta = 181,
  rho = 2 * round(sqrt(sum((dim(image))^2))/2) + 1,
  fn = "polar"
)

nfft_inv_radon(image, n, iter = 10, fn = "polar")
}
\arguments{
\item{image}{square image}

\item{theta}{Number of theta slices}

\item{rho}{Number of rho slices}

\item{fn}{Whether to use polar or linotype (polar by default)}

\item{n}{size of image}

\item{iter}{number of iterations for inverse}
}
\description{
A radon transformation using the non-equispaced
fast Fourier transform (NFFT).
Requires a square image.
}
\section{Functions}{
\itemize{
\item \code{nfft_inv_radon}: Inverse Radon Transform using NFFT
}}

\examples{
P <- PET::phantom()
image(P)
P_radon <- nfft_radon(P, 514, 514, fn = "polar")
image(P_radon)
P_inv <- nfft_inv_radon((P_radon), n = 257, iter = 2, fn = "polar")
image(P_inv)
}
