% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phantom_example}
\alias{phantom_example}
\title{PET Phantom Example}
\format{
A \eqn{257 \times 257}{257 x 257} matrix representing a
PET phantom. From the output of \link[PET:phantom]{PET::phantom}.
}
\usage{
phantom_example
}
\description{
PET Phantom Example
}
\keyword{datasets}
