% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNFFT.R
\docType{package}
\name{rNFFT}
\alias{rNFFT}
\title{R wrapper for the NFFT library (nonequispaced nodes Fast Fourier Transform)}
\description{
Provides a wrapper around the NFFT 3.0 C library, which implements
the fast Fourier transform for non-equispaced nodes, its adjoint, and its
inversion. This provides functions for the NFFT and Solver routines.
}
