% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNFFT.R
\name{test}
\alias{test}
\title{Test Function}
\usage{
test(M, N)
}
\arguments{
\item{M}{integer, number of nodes}

\item{N}{integer}
}
\value{
Null
}
\description{
Test Function
}
\details{
Generates Test Output
}
\examples{

test(19L,14L)

}
\author{
Geoffrey Z. Thompson
}
