% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNFFT.R
\name{test_nfft}
\alias{test_nfft}
\alias{test_solve}
\alias{nfft_test_2d}
\title{Test Function}
\usage{
test_nfft(m, n)

test_solve(m, n, iterations)

nfft_test_2d()
}
\arguments{
\item{m}{integer, number of nodes}

\item{n}{integer, number of frequences (even, less than \code{m}).}

\item{iterations}{Number of iterations (for solver)}
}
\value{
TRUE
}
\description{
Test Function
}
\details{
Generates Test Output specified in NFFT documentation and source code.
Compare the output to the results of the examples.
}
\section{Functions}{
\itemize{
\item \code{test_solve}: Test solver: displays the 1D solver.

\item \code{nfft_test_2d}: 2D test: displays the 2D test (no input)
}}

\examples{
set.seed(20190722)
test_nfft(19L,14L)
test_solve(8L,16L,9L)
set.seed(20190728)
nfft_test_2d()
}
\author{
Geoffrey Z. Thompson
}
