% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNFFT.R
\name{nfft_radon}
\alias{nfft_radon}
\alias{nfft_inv_radon}
\title{Radon transform using NFFT}
\usage{
nfft_radon(image, Theta = 181, Rho = 2 *
  round(sqrt(sum((dim(image))^2))/2) + 1, fn = "polar")

nfft_inv_radon(image, N, iter = 10, fn = "polar")
}
\arguments{
\item{image}{square image}

\item{Theta}{Number of theta slices}

\item{Rho}{Number of rho slices}

\item{fn}{Whether to use polar or linotype (polar by default)}

\item{N}{size of image}

\item{iter}{number of iterations for inverse}
}
\description{
Radon transform using NFFT
}
\section{Functions}{
\itemize{
\item \code{nfft_inv_radon}: Inverse Radon Transform using NFFT
}}

