% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNFFT.R
\name{nfft_solver_1d}
\alias{nfft_solver_1d}
\title{Inverse of the 1-D NFFT.}
\usage{
nfft_solver_1d(x, f, n, eps = 1e-12, iterations = 10)
}
\arguments{
\item{x}{input vector (real, length \code{m})}

\item{f}{Outputted Fourier coefficients (complex, length \code{m})}

\item{n}{Number of coefficients for the inverse - must be even.}

\item{eps}{convergence criterion}

\item{iterations}{number of iterations for solve to run (at most)}
}
\value{
vector of length \code{n} of solutions \eqn{\hat{f}}{f_hat}.
}
\description{
1-D NFFT solver
}
\details{
Unlike in the case of the equispaced Fourier transform, the adjoint is NOT
the inverse of the non-equispaced Fourier transform. The solution must be found
numerically.
}
\examples{
set.seed(20190722)
x = runif(8) - 0.5
f = 1:8
for(i in 1:8) f[i] = runif(1)*1i + runif(1)
x
f
nfft_solver_1d(x,f,16,1e-35,8)
nfft_1d(x, nfft_solver_1d(x,f,16,eps = 1e-35))
}
\author{
Geoffrey Z. Thompson
}
