% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNFFT.R
\name{test_nfft}
\alias{test_nfft}
\title{Test Function}
\usage{
test_nfft(M, N)
}
\arguments{
\item{M}{integer, number of nodes}

\item{N}{integer, number of frequences (even, less than \code{M}).}
}
\value{
TRUE
}
\description{
Test Function
}
\details{
Generates Test Output. Compare the output the results
}
\examples{
set.seed(20190722)
test_nfft(19L,14L)

}
\author{
Geoffrey Z. Thompson
}
