% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNFFT.R
\name{test_nfft}
\alias{test_nfft}
\alias{test_solve}
\alias{nfft_test_2d}
\title{Test Function}
\usage{
test_nfft(M, N)

test_solve(M, N, iterations)

nfft_test_2d()
}
\arguments{
\item{M}{integer, number of nodes}

\item{N}{integer, number of frequences (even, less than \code{M}).}

\item{iterations}{Number of iterations (for solver)}
}
\value{
TRUE
}
\description{
Test Function
}
\details{
Generates Test Output. Compare the output the results
}
\section{Functions}{
\itemize{
\item \code{test_solve}: Test solver: displays the 1D solver.

\item \code{nfft_test_2d}: 2D test: displays the 2D test (no input)
}}

\examples{
set.seed(20190722)
test_nfft(19L,14L)
test_solve(8L,16L,9L)
set.seed(20190728)
nfft_test_2d()
}
\author{
Geoffrey Z. Thompson
}
